Factor out dir traversal, move allocs to heap
Factor out directory traversal to separate function
Move allocations to heap. I have found that we are less likely to hit
stack bugs when doing large allocations on the heap rather then the stack.
Free'd memory, not strickly necessary since it's in the main function
but easier to find other leaks with valgrind this way.